+++ VerizonAds/src/main/java/com/applovin/mediation/adapters/VerizonAdsMediationAdapter.java
@@ -52,6 +52,7 @@ import com.yahoo.ads.yahoonativecontroller.NativeImageComponent;
 import com.yahoo.ads.yahoonativecontroller.NativeTextComponent;
 import com.yahoo.ads.yahoonativecontroller.NativeVideoComponent;
 
+import java.lang.ref.WeakReference;
 import java.lang.reflect.Method;
 import java.util.Collections;
 import java.util.HashMap;
@@ -290,7 +291,6 @@ public class VerizonAdsMediationAdapter
         RequestMetadata requestMetadata = createRequestMetadata( parameters.getBidResponse() );
         final InlinePlacementConfig config = new InlinePlacementConfig( placementId, requestMetadata, Collections.singletonList( adSize ) );
 
-
         ActivityStateManager activityStateManager = YASAds.getActivityStateManager();
         activityStateManager.setState( activity, ActivityStateManager.ActivityState.RESUMED );
 
@@ -318,9 +318,20 @@ public class VerizonAdsMediationAdapter
         updatePrivacyStates( parameters );
         updateLocationCollectionEnabled( parameters );
 
+        // Yahoo requires us to pass them an activity for their viewability and impression tracking.
+        // Note: if the provided activity is different from the final activity that the native ad view is attached,
+        // Yahoo may fail to fire impression tracking. The pub would need to load and attach the native ad with the same activity.
+        if ( activity == null )
+        {
+            e( "Native ad (" + placementId + ") failed to load: activity reference is null..." );
+            listener.onNativeAdLoadFailed( MaxAdapterError.INVALID_LOAD_STATE );
+
+            return;
+        }
+
         Context context = getContext( activity );
 
-        NativeAdListener nativeAdListener = new NativeAdListener( parameters.getServerParameters(), context, listener );
+        NativeAdListener nativeAdListener = new NativeAdListener( parameters.getServerParameters(), activity, listener );
         nativeAd = new NativeAd( context, placementId, nativeAdListener );
 
         RequestMetadata requestMetadata = createRequestMetadata( bidResponse );
@@ -763,13 +774,13 @@ public class VerizonAdsMediationAdapter
             implements NativeAd.NativeAdListener
     {
         private final Bundle                     serverParameters;
-        private final Context                    context;
+        private final WeakReference<Activity>    activityRef;
         private final MaxNativeAdAdapterListener listener;
 
-        private NativeAdListener(final Bundle serverParameters, final Context context, final MaxNativeAdAdapterListener listener)
+        private NativeAdListener(final Bundle serverParameters, final Activity activity, final MaxNativeAdAdapterListener listener)
         {
             this.serverParameters = serverParameters;
-            this.context = context;
+            this.activityRef = new WeakReference<>( activity );
             this.listener = listener;
         }
 
@@ -800,14 +811,22 @@ public class VerizonAdsMediationAdapter
                     NativeTextComponent ctaComponent = (NativeTextComponent) nativeAd.getComponent( "callToAction" );
                     if ( ctaComponent != null ) callToAction = ctaComponent.getText();
 
+                    final Activity activity = activityRef.get();
+                    if ( activity == null )
+                    {
+                        e( "Native ad (" + nativeAd + ") failed to load: activity reference is null when ad is loaded" );
+                        listener.onNativeAdLoadFailed( MaxAdapterError.INVALID_LOAD_STATE );
+
+                        return;
+                    }
+
                     // NOTE: Yahoo's SDK only returns ImageView with the image pre-cached, we cannot use the 'getUri()' getter
                     // since it is un-cached and our SDK will attempt to re-cache it, and we do not support passing ImageView for custom native
-
                     MaxNativeAd.MaxNativeAdImage iconImage = null;
                     NativeImageComponent iconComponent = (NativeImageComponent) nativeAd.getComponent( "iconImage" );
                     if ( iconComponent != null )
                     {
-                        ImageView iconView = new ImageView( context );
+                        ImageView iconView = new ImageView( activity );
                         iconComponent.prepareView( iconView );
 
                         Drawable drawable = iconView.getDrawable();
@@ -828,14 +847,14 @@ public class VerizonAdsMediationAdapter
                     {
                         mediaViewAspectRatio = (float) nativeVideoComponent.getWidth() / (float) nativeVideoComponent.getHeight();
 
-                        mediaView = new VideoPlayerView( context );
+                        mediaView = new VideoPlayerView( activity );
                         nativeVideoComponent.prepareView( (VideoPlayerView) mediaView );
                     }
                     else if ( nativeImageComponent != null )
                     {
                         mediaViewAspectRatio = (float) nativeImageComponent.getWidth() / (float) nativeImageComponent.getHeight();
 
-                        mediaView = new ImageView( context );
+                        mediaView = new ImageView( activity );
                         nativeImageComponent.prepareView( (ImageView) mediaView );
 
                         Drawable drawable = ( (ImageView) mediaView ).getDrawable();
@@ -876,7 +895,7 @@ public class VerizonAdsMediationAdapter
                         builder.setMediaContentAspectRatio( mediaViewAspectRatio );
                     }
 
-                    MaxNativeAd maxNativeAd = new MaxYahooNativeAd( builder );
+                    MaxNativeAd maxNativeAd = new MaxYahooNativeAd( activity, builder );
 
                     CreativeInfo creativeInfo = nativeAd.getCreativeInfo();
                     Bundle extraInfo = new Bundle( 1 );
@@ -940,18 +959,29 @@ public class VerizonAdsMediationAdapter
     private class MaxYahooNativeAd
             extends MaxNativeAd
     {
+        private final WeakReference<Activity> activityRef;
 
-        private MaxYahooNativeAd(final Builder builder)
+        private MaxYahooNativeAd(final Activity activity, final Builder builder)
         {
             super( builder );
+
+            this.activityRef = new WeakReference<>( activity );
         }
 
         @Override
         public void prepareViewForInteraction(final MaxNativeAdView maxNativeAdView)
         {
+            final NativeAd nativeAd = VerizonAdsMediationAdapter.this.nativeAd;
             if ( nativeAd == null )
             {
-                e( "Failed to register native ad view for interaction. Native ad is null" );
+                e( "Failed to register native ad view for interaction. Native ad is null." );
+                return;
+            }
+
+            final Activity activity = activityRef.get();
+            if ( activity == null )
+            {
+                e( "Native ad (" + nativeAd + ") failed to prepare native view for interaction. Activity reference is null." );
                 return;
             }
 
@@ -986,7 +1016,7 @@ public class VerizonAdsMediationAdapter
                 iconComponent.prepareView( maxNativeAdView.getIconImageView() );
             }
 
-            nativeAd.registerContainerView( maxNativeAdView );
+            nativeAd.registerContainerView( maxNativeAdView, activity );
         }
     }